foo/requirements.txt
polygon-api-client
python-dotenv
yfinance